+++ jana/__init__.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 # Copyright (C) 2026 Tonic
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"
 __title__ = "Jana"